+++ src/Lawn/Coin.cpp
@@ -688,7 +688,7 @@ void Coin::UpdateCollected()
     {
         if (mCollectionDistance < 15.0f)
         {
-            if (!mBoard->mHelpDisplayed[static_cast<int>(AdviceType::ADVICE_UNLOCKED_MODE)])
+            if (!mBoard->mHelpDisplayed[AdviceType::ADVICE_UNLOCKED_MODE])
             {
                 if (mType == CoinType::COIN_PRESENT_MINIGAMES)
                 {
@@ -1141,14 +1141,14 @@ void Coin::Collect()
         mBoard->mChocolateCollected++;
         mApp->AddTodParticle(mPosX + 30.0f, mPosY + 30.0f, mRenderOrder + 1, ParticleEffect::PARTICLE_PRESENT_PICKUP);
 
-        if (mApp->mPlayerInfo->mPurchases[static_cast<int>(StoreItem::STORE_ITEM_CHOCOLATE)] < PURCHASE_COUNT_OFFSET)
+        if (mApp->mPlayerInfo->mPurchases[StoreItem::STORE_ITEM_CHOCOLATE] < PURCHASE_COUNT_OFFSET)
         {
             mBoard->DisplayAdvice("[ADVICE_FOUND_CHOCOLATE]", MessageStyle::MESSAGE_STYLE_HINT_TALL_FAST, AdviceType::ADVICE_NONE);
-            mApp->mPlayerInfo->mPurchases[static_cast<int>(StoreItem::STORE_ITEM_CHOCOLATE)] = PURCHASE_COUNT_OFFSET + 1;
+            mApp->mPlayerInfo->mPurchases[StoreItem::STORE_ITEM_CHOCOLATE] = PURCHASE_COUNT_OFFSET + 1;
         }
         else
         {
-            mApp->mPlayerInfo->mPurchases[static_cast<int>(StoreItem::STORE_ITEM_CHOCOLATE)]++;
+            mApp->mPlayerInfo->mPurchases[StoreItem::STORE_ITEM_CHOCOLATE]++;
         }
 
         mDisappearCounter = 0;
