+++ src/Lawn/CursorObject.cpp
@@ -77,7 +77,7 @@ void CursorObject::Draw(Graphics* g)
         break;
 
     case CursorType::CURSOR_TYPE_WATERING_CAN:
-        if (mApp->mPlayerInfo->mPurchases[static_cast<int>(StoreItem::STORE_ITEM_GOLD_WATERINGCAN)])
+        if (mApp->mPlayerInfo->mPurchases[StoreItem::STORE_ITEM_GOLD_WATERINGCAN])
         {
             g->DrawImage(IMAGE_ZEN_GOLDTOOLRETICLE, -62, -37);
             g->DrawImage(IMAGE_WATERINGCANGOLD, -3, 12);
